updated to wxWidgets 3.2.2.1, added fix for MacOs HiRes blurry resources

+++ src/MimeTypeListFrame.cpp
@@ -74,7 +74,7 @@ void MimeTypeListFrame::OnListItemActivated( wxListEvent& event )
 {
   wxString key;
   wxFileName fname;
-  int index = event.GetIndex();
+  long index = event.GetIndex();
 
   EditEntryDialog dlg(this);
